refactor: update hashPurpose filename
