+++ libs/common/src/auth/login-strategies/password-login.strategy.spec.ts
@@ -7,7 +7,7 @@ import { LogService } from "../../abstractions/log.service";
 import { MessagingService } from "../../abstractions/messaging.service";
 import { PlatformUtilsService } from "../../abstractions/platformUtils.service";
 import { StateService } from "../../abstractions/state.service";
-import { HashPurpose } from "../../enums/hashPurpose";
+import { HashPurpose } from "../../enums/hash-purpose";
 import { Utils } from "../../misc/utils";
 import { SymmetricCryptoKey } from "../../models/domain/symmetric-crypto-key";
 import { AuthService } from "../abstractions/auth.service";
