+++ libs/common/src/enums/hash-purpose.ts
